vue-fastapi/backend/requirements.txt
requests
fastapi
python-dotenv
